+++ adapter/cli/src/main/kotlin/CliMain.kt
@@ -1,6 +1,8 @@
 import adapters.primary.CliAdapter
+import application.usecases.TokenizeUseCase
 
 fun main(args: Array<String>) {
-    val cliAdapter = CliAdapter()
+    val tokenizeUseCase = TokenizeUseCase()
+    val cliAdapter = CliAdapter(tokenizeUseCase)
     cliAdapter.handleCommand(args)
 }